requirements.txt
azure-core
azure-search-documents
flask
langchain
langchain-core
langchain-openai
langgraph
python-dotenv
